feat: Add lane following function
It's currently very jerky and and does not make any decisions, but this is an initial version of a lane following function using the s and d coordinates of the car.

The next step is to use the map waypoints to calculate a spline that can then be used to create the path vector. This should hopefully smooth things out a bit.

I also need to figure out how to accelerate and decelerate correctly to prevent jerk at the beginning.