feat: Complete project
All major requirements for the project have been satisfied. I still have to do the write up but that's it.

There are a few bugs still, like if a car stops suddenly (which in theory shouldn't happen because it's a highway but I did see it happen once and technically it is a possibility), that will most likely result in a collision. However, that is definitely an edge case and also not unrealistic so I'm not going to worry about it for now especially since it wasn't really within the scope of the project.

The parameters could still me tweaked a little more like following distance and speed but they seem to work well enough for now.